Refactor the code to accomodate the overall architecture.
Effectively, the code now allows for new filetypes (and also new features,
adapters, etc. to be added easily). Good code practices with package design,
the use of __init__.py, and relative imports have been followed.

It now remains to integrate the dataset preparation and ML scripts with
those for the adapters and feature collection.