Updated Pip requirements (explained below). Updated mysql python installation. Updated APT requirments    PIP CHANGES ** Pip Updates
enum34 was updated to the latest (only increased minor version)

ipaddress was updated to the latest (only increased hotfix)

requests was updated to latest, (only increased minor version)

chardet was upgraded to 3.0.4. Although this is a major version change,
after researching API changes I found no breaking changes.
It appears that 3.0.4 is only faster / more accurate than 2.3.
The version of requests we upgraded to requires this version of chardet.

cryptography was upgraded to 2.9.2. This is because packages which depended
on this were also upgraded.
The frontend code uses this library but changelogs revealed no major
API changes to used functions, and thus code should be compatible.

M2Crypto was updated to support latest openssl (only minor version update)

** Pip Removals / Replacements

mysqlclient replaces mysql-python. Mysqlclient is designed to be a maintained
version of the same code as mysql-python, and thus will only be
more stable.

Django was removed as it is not necessary for the computational nodes
(it is only necessary for the webportal)

gglsbl has been removed, as this functionality is now done by the
microservice

wakeonlan is only used by the physical machine type

Woocommerce was removed as we no longer use woocommerce for tracking purchases.

** Pip Additions

Yara python was added to replace the apt package which is no longer available.

Click was added, as a specific version is required by HTTPReplay (but not flask,
which also uses it).